FORMS-3217 Pasted component have unique name
