+++ ui.af.apps/src/main/content/jcr_root/apps/core/fd/components/form/container/v2/container/clientlibs/editorhook/js/copypastehook.js
@@ -24,11 +24,10 @@
      * @param config
      */
     author.persistence.PostRequest.prototype.prepareCopyParagraph = function (config) {
-        var request = _superPrepareCopyParagraph.apply(this, [config]),
-            nodeNameToCopy = config.path.substring(config.path.lastIndexOf("/") + 1);
+        var request = _superPrepareCopyParagraph.apply(this, [config]);
         if (config.parentPath 
             && author.editables.find(config.parentPath)[0].dom.closest(FORM_CONTAINER_SELECTOR).length > 0) {
-            var uniqueName = getUniqueName(config.parentPath, nodeNameToCopy);
+            var uniqueName = getUniqueName(config.parentPath, config.path);
             return (
                 this.setParam("./name", uniqueName)
                     .setParam("./dataRef@Delete", "deleted value")
@@ -40,8 +39,8 @@
         }
     };
 
-    getUniqueName = function (path, nodeNameToCopy) {
-        var resourceExistSelector = ".fdResourceExists.json?childNodeName=" + nodeNameToCopy,
+    getUniqueName = function (path, nodePathToCopy) {
+        var resourceExistSelector = ".fdResourceExists.json?nodePathToCopy=" + nodePathToCopy,
             url = path + resourceExistSelector,
             response = CQ.shared.HTTP.get(url);
         return JSON.parse(response.responseText).uniqueName;
